% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{absolute_to_relative}
\alias{absolute_to_relative}
\title{Convert Absolute Returns}
\usage{
absolute_to_relative(r_n, n)
}
\arguments{
\item{r_n}{an absolute return over a single time period}

\item{n}{the number of time periods over which to compute the
logarithmic return}
}
\value{
a relative (log) percent return
}
\description{
Convert absolute returns to relative (log) returns.
}
