% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolios.R
\name{change_amounts}
\alias{change_amounts}
\title{Update a portfolio from a named list of new_amounts}
\usage{
change_amounts(portfolio, changed_amounts, nonneg = TRUE)
}
\description{
Update a portfolio from a named list of new_amounts
}
