% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocations.R
\name{check_allocation}
\alias{check_allocation}
\title{Check that allocation percents sum to 1.0}
\usage{
check_allocation(percents)
}
\arguments{
\item{percents}{a vector of percents to check}
}
\value{
a boolean indicating if the percents in fact do sum to 1.0
}
\description{
Check that allocation percents sum to 1.0
}
