% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{find_begin}
\alias{find_begin}
\title{Find a Date in a List}
\usage{
find_begin(dates, d, right = FALSE)
}
\arguments{
\item{dates}{a list of dates}

\item{d}{the reference date}

\item{right}{whether to check from the right of the given date}
}
\value{
the date found
}
\description{
The first date in the list before the given date
}
