% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{find_begin_position}
\alias{find_begin_position}
\title{The position of the first date in the list before the given date}
\usage{
find_begin_position(dates, d, right = FALSE)
}
\description{
The position of the first date in the list before the given date
}
