% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{find_end_position}
\alias{find_end_position}
\title{Find a Date's Position in a List}
\usage{
find_end_position(dates, d, right = FALSE)
}
\arguments{
\item{dates}{a list of dates}

\item{d}{the reference date}

\item{right}{whether to check from the right of the given date}
}
\value{
the position of the date found
}
\description{
The position of the first date in the list after the given date
}
