% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{make_bootstrap_historical}
\alias{make_bootstrap_historical}
\title{Make a bootstrap-sampled series of returns}
\usage{
make_bootstrap_historical(ts, block_size = 5)
}
\arguments{
\item{ts}{an \code{xts} time-series of absolute returns}
}
\description{
Creates a time-series of returns sampled from historical data, with
the sampling respecting time dependence.
}
