% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{make_empty_model}
\alias{make_empty_model}
\title{An empty model}
\usage{
make_empty_model(portfolio, allocations)
}
\arguments{
\item{portfolio}{a \code{portfolio} object}

\item{allocations}{an \code{allocation} object}
}
\description{
Creates a model containing only a portfolio and an allocation.
Useful for setting up subsimulations with varying transaction paths
and dates.
}
