% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_geom_block_sample}
\alias{make_geom_block_sample}
\title{Stationary bootstrap with geometric block size}
\usage{
make_geom_block_sample(ts, block_size, R)
}
\arguments{
\item{ts}{a time series to sample from}

\item{block_size}{is mean block size for sampling}

\item{R}{the number of samples}
}
\description{
A function for sampling time-dependent data.
}
\details{
source: https://eranraviv.com/bootstrapping-time-series-r-code/
}
