% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{make_historical}
\alias{make_historical}
\title{Make an asset from a time series of absolute (ordinary) returns}
\usage{
make_historical(ts)
}
\arguments{
\item{ts}{an \code{xts} time-series of absolute returns}
}
\value{
an \code{asset} object for the time-series
}
\description{
Make an asset from a time series of absolute (ordinary) returns
}
