% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocations.R
\name{make_linear_allocation_path}
\alias{make_linear_allocation_path}
\title{Create an allocation path with linear interpolation}
\usage{
make_linear_allocation_path(asset_names, change_dates, asset_percents)
}
\arguments{
\item{asset_names}{a vector of strings identifying each holding}

\item{change_dates}{a vector of dates indicating when to change
target allocations; these function as the "knots" of the
interpolation}

\item{asset_percents}{a list of vectors of numbers between 0 and 1
identifying the percentage desired of each holding. See above
for structure.}
}
\value{
an allocation path object#' @param asset_names names of
    holdings
}
\description{
Give a list of n asset_names, a list of k dates, and k+1 lists of n
percents, will return an allocation function that changes by linear
interpolation from one percent to the next on the specified dates.
If a date is given before the first change or after the last
change, it returns a constant allocation.
}
