% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{make_model}
\alias{make_model}
\title{Create a model object}
\usage{
make_model(portfolio, allocations, transactions, dates)
}
\arguments{
\item{portfolio}{A \code{portfolio} object}

\item{allocations}{An \code{allocation} object}

\item{transactions}{A \code{transaction} object}

\item{dates}{A vector of dates over which the model should be
simulated.}
}
\description{
Creates a model object for use in financial paths.
}
