% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial-path.R
\name{make_path}
\alias{make_path}
\title{Create a financial path}
\usage{
make_path(model, nonneg = TRUE, verbose = FALSE)
}
\arguments{
\item{model}{a named list containing \code{transactions},
\code{allocations}, \code{portfolio}, \code{asset_names}, and
\code{dates}}

\item{nonneg}{whether to disallow a negative balance. Setting this
to \code{FALSE} (allowing a negative balance) can be useful for
understanding shortfalls.}

\item{verbose}{whether to output diagnostic information. Currently unused.}
}
\value{
a financial path, which is an \code{xts} timeseries
    containing balances for each holding and the total balance, as
    well as any transactions that occured and the investment return
    as a percent change.
}
\description{
Evaluates the model given in \code{model} by stepping through the
given dates and applying transformations as follows: adjust for
returns over the preceeding period, apply contributions or
withdrawals (transactions), rebalance according to the given
allocations.
}
\examples{
library(tidyverse)
library(zoo)
library(lubridate)
## Create a model and financial path

# Time series of returns
data(simreturns)

# Historical assets
simstock_asset <- make_historical(simreturns)
simbond_asset <- make_historical(simreturns)
dates <- as_datetime(index(simreturns))

# Portfolio with S&P 500 and 10-year T-bonds. Yearly transaction
# of $1000. Linear allocation.
holdings <- c("stocks", "bonds")
port <- make_portfolio(holdings,
                       c(simstock_asset,
                         simbond_asset),
                         c(0,0))
alloc <- make_linear_allocation_path(holdings,
                                    c(ymd("1990-01-01"), ymd("2000-01-01")),
                                    list(c(0.9, 0.1), c(0.6, 0.4)))
trans <- make_transactions_on_dates(rep(1000, length(dates)),
                                    dates)
model <- make_model(port, alloc, trans, dates)

# Evaluate the model
path <- make_path(model)
}
