% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolios.R
\name{make_portfolio}
\alias{make_portfolio}
\alias{get_assets}
\alias{get_amounts}
\alias{get_total}
\title{Create a portfolio}
\usage{
make_portfolio(asset_names, assets, amounts)

get_assets(portfolio)

get_amounts(portfolio)

get_total(portfolio)
}
\arguments{
\item{asset_names}{a vector of strings identifying each holding}

\item{assets}{a vector of \code{asset} objects}

\item{amounts}{a vector of the dollar amount held for each holding}

\item{portfolio}{the portfolio to return information from}
}
\value{
a new \code{portfolio} object
}
\description{
Creates a portfolio of holdings and the amount held.
}
