% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{make_relative_historical}
\alias{make_relative_historical}
\title{Make an asset from a time series of relative (log) returns}
\usage{
make_relative_historical(ts)
}
\arguments{
\item{ts}{an \code{xts} time-series of relative returns}
}
\value{
an \code{asset} object for the time series
}
\description{
Make an asset from a time series of relative (log) returns
}
