% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{make_relative_returns}
\alias{make_relative_returns}
\title{Compute relative returns of an asset}
\usage{
make_relative_returns(x)
}
\arguments{
\item{x}{a vector of prices}
}
\value{
a vector of relative (log) returns
}
\description{
Computes the relative returns of an asset given its prices over
some regular period, that is \eqn{ \log{\frac{x_k}{x_{k-1}}}}.
}
