% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{make_returns}
\alias{make_returns}
\title{Assets
------
An Asset has type: Time -> Time -> Return}
\usage{
make_returns(x)
}
\arguments{
\item{x}{A time-series of prices}
}
\description{
An asset is identified with an accumulator. It takes an increment and
returns the asset's percent change over that increment.
}
\details{
An asset should in addition have as an attribute 'dates' the vector
of dates over which it is defined. If left as 'NULL' it is assumed
to be valid for any date.
Compute absolute returns of an asset

Computes the absolute returns of an asset given its prices over
some regular period, that is \eqn{ \frac{x_k - x_{k-1}}{x_{k-1}}}.
}
