% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{make_transactions_on_dates}
\alias{make_transactions_on_dates}
\title{Create a transaction path}
\usage{
make_transactions_on_dates(amounts, dates)
}
\arguments{
\item{amounts}{a vector of dollar amounts; positive numbers are
contributions and negative numbers are withdrawals}

\item{dates}{a vector of the dates on which each transaction occurs}
}
\description{
Creates a transaction path the applies each transaction on a given date.
}
