% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{note_return}
\alias{note_return}
\title{Compute Returns from Treasury Notes}
\usage{
note_return(r1, r2)
}
\arguments{
\item{r1}{starting rate}

\item{r2}{ending rate}
}
\value{
the percent change in the price of the note
}
\description{
From the interest rates on a 10-year Treasury note Computes the
change in price of the note due to the change in interest rate,
based on the present-value valuation.
}
