% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{proportion_elapsed}
\alias{proportion_elapsed}
\title{The proportion of time elapsed to current in the interval (start, end)}
\usage{
proportion_elapsed(start, current, end)
}
\description{
The proportion of time elapsed to current in the interval (start, end)
}
