% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{relative_to_absolute}
\alias{relative_to_absolute}
\title{Convert Relative Returns}
\usage{
relative_to_absolute(r_log, n)
}
\arguments{
\item{r_log}{a logarithmic (relative) return}

\item{n}{the number of time periods over which to compute the
absolute return.}
}
\description{
Convert relative returns to absolute returns.
}
