% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simreturns}
\alias{simreturns}
\title{Simulated Financial Returns}
\format{A data frame with 91 rows and 2 variables:
\describe{
  \item{Stock.Returns}{fictional yearly stock percent returns from 1928 to 2018}
  \item{Bond.Returns}{fictional yearly bond percent returns from 1928 to 2918}
}}
\usage{
simreturns
}
\description{
This data set contains two series of fictional returns from the years 1928
to 2018. One series is meant to represent a stock and the other is meant
to represent a bond.
}
\keyword{datasets}
