% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{find_end_position}
\alias{find_end_position}
\title{The position of the first date in the list after the given date}
\usage{
find_end_position(dates, d, right = FALSE)
}
\description{
The position of the first date in the list after the given date
}
