% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocations.R
\name{make_constant_allocation_path}
\alias{make_constant_allocation_path}
\title{Create a constant allocation path}
\usage{
make_constant_allocation_path(asset_names, percents)
}
\description{
The same allocation, forever.
}
