% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocations.R
\name{make_linear_allocation_path}
\alias{make_linear_allocation_path}
\title{Create an allocation path with linear interpolation}
\usage{
make_linear_allocation_path(asset_names, change_dates, asset_percents)
}
\description{
Give a list of n asset_names, a list of k dates, and k+1 lists of n
percents, will return an allocation function that changes by linear
interpolation from one percent to the next on the specified dates.
If a date is given before the first change or after the last
change, it returns a constant allocation.
}
