% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{proportion_elapsed}
\alias{proportion_elapsed}
\title{Find the Proportion of Elapsed Time}
\usage{
proportion_elapsed(start, current, end)
}
\arguments{
\item{start}{the stating date}

\item{current}{the current date}

\item{end}{the ending date}
}
\value{
a proportion
}
\description{
The proportion of time elapsed to current in the interval (start, end)
}
