% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolios.R
\name{rebalance}
\alias{rebalance}
\title{Rebalance a portfolio}
\usage{
rebalance(portfolio, allocation, transaction, nonneg = TRUE)
}
\description{
Rebalance a portfolio according to an allocation while applying any
additional transactions.
}
